--- kidney_transplant_llm/pydantic_study_variables.py
@@ -1,917 +1,0 @@
-import json
-import os
-
-from enum import StrEnum, auto
-from pydantic import BaseModel, Field
-
-
-class SpanAugmentedMention(BaseModel):
-    has_mention: bool = Field(
-        False,
-        description="Whether there is any mention of this variable in the text."
-    )
-    spans: list[str] = Field(
-        default_factory=list,
-        description="The text spans where this variable is mentioned."
-    )
-
-
-###############################################################################
-# History of Multiple Transplants
-#
-# Mentions relevant in tracking if this patient has a history of multiple transplants,
-# renal or otherwise.
-###############################################################################
-class MultipleTransplantHistoryMention(SpanAugmentedMention):
-    """
-    Does this patient have a history of multiple transplants, renal or otherwise?
-    For use in reevaluating the patients in our cohort, excluding patients with a history
-    of multiple transplants from our analysis.
-    """
-    multiple_transplant_history: bool = Field(
-        False,
-        description="Whether there is any mention of a history of multiple transplants. "
-    )
-
-###############################################################################
-# Donor Characteristics
-#
-# For a given transplant, these should be static over time
-###############################################################################
-
-# Dates are treated as strings - no enum needed
-class DonorTransplantDateMention(SpanAugmentedMention):
-    """
-    Date of the first renal transplant. If there is only one transplant mentioned,
-    assume that this is the first transplant. If a POD (post-operative day) is mentioned,
-    use that to infer the date.
-    """
-    donor_transplant_date: str | None = Field(
-        None,
-        description='Exact date of first renal transplant; use YYYY-MM-DD format in your response. Only highlight date mentions with an explicit day, month, and year (e.g. 2020-01-15). All other date mentions, or an absence of a date mention, should be indicated with None.'
-    )
-
-
-class DonorType(StrEnum):
-    LIVING = 'Donor was alive at time of renal transplant'
-    DECEASED = 'Donor was deceased at time of renal transplant'
-    NOT_MENTIONED = "Donor was not mentioned as living or deceased"
-
-
-class DonorTypeMention(SpanAugmentedMention):
-    """
-    Type of donor in the first renal transplant. If there is only one transplant
-    mentioned, assume that this is the first transplant. Exclude donors that are only
-    hypothetical or under evaluation/assessment.
-    """
-    donor_type: DonorType = Field(
-        DonorType.NOT_MENTIONED,
-        description='Was the first renal donor living at the time of renal transplant?'
-    )
-
-
-class DonorRelationship(StrEnum):
-    RELATED = 'Donor was biologically related to the renal transplant recipient'
-    UNRELATED = 'Donor was biologically unrelated to the renal transplant recipient'
-    NOT_MENTIONED = "Donor relationship status was not mentioned"
-
-
-class DonorRelationshipMention(SpanAugmentedMention):
-    """
-    Relatedness of the donor and recipient in the first renal transplant. If there is only one
-    transplant mentioned, assume that this is the first transplant. Exclude donors that are only
-    hypothetical or under evaluation/assessment.
-    """
-    donor_relationship: DonorRelationship = Field(
-        DonorRelationship.NOT_MENTIONED,
-        description='Was the first renal transplant donor biologically related to the recipient?'
-    )
-
-
-class DonorHlaMatchQuality(StrEnum):
-    WELL = "Well matched (0-1 mismatches) OR recipient explicitly documented as not sensitized"
-    MODERATE = "Moderately matched (2-4 mismatches) OR recipient explicitly documented as sensitized"
-    POOR = "Poorly matched (5-6 mismatches) OR recipient explicitly documented as highly sensitized"
-    NOT_MENTIONED = "HLA match quality not mentioned"
-
-
-class DonorHlaMatchQualityMention(SpanAugmentedMention):
-    """
-    Human leukocyte antigen (HLA) match quality of the first renal transplant. If there is only one
-    transplant mentioned, assume that this is the first transplant.
-    """
-    donor_hla_match_quality: DonorHlaMatchQuality = Field(
-        DonorHlaMatchQuality.NOT_MENTIONED,
-        description='What was the HLA match quality for the first renal transplant?'
-    )
-
-
-class DonorHlaMismatchCount(StrEnum):
-    ZERO = "0"
-    ONE = "1"
-    TWO = "2"
-    THREE = "3"
-    FOUR = "4"
-    FIVE = "5"
-    SIX = "6"
-    NOT_MENTIONED = 'HLA mismatch count not mentioned'
-
-
-class DonorHlaMismatchCountMention(SpanAugmentedMention):
-    """
-    Human leukocyte antigen (HLA) mismatch count for the first renal transplant. If there is only
-    one transplant mentioned, assume that this is the first transplant.
-    """
-    donor_hla_mismatch_count: DonorHlaMismatchCount = Field(
-        DonorHlaMismatchCount.NOT_MENTIONED,
-        description='What was the donor-recipient HLA mismatch count for the first renal transplant?'
-    )
-
-class Serostatus(StrEnum):
-    """
-    Serostatus classification based on IgG serology or explicit seropositive/seronegative statements.
-    - SEROPOSITIVE: Documented IgG positive / seropositive
-    - SERONEGATIVE: Documented IgG negative / seronegative
-    - NOT_MENTIONED: No serostatus documentation found
-    """
-    SEROPOSITIVE = "SEROPOSITIVE"     
-    SERONEGATIVE = "SERONEGATIVE"     
-    NOT_MENTIONED = "NOT_MENTIONED"   
-
-class SerostatusDonorMention(SpanAugmentedMention):
-    """
-    Overall serostatus of the donor at the time of first renal transplant.
-    """
-    serostatus: Serostatus = Field(
-        Serostatus.NOT_MENTIONED,
-        description=(
-            "Overall serostatus of the renal donor for ANY virus. "
-            "Set SEROPOSITIVE if the note documents the donor as seropositive for "
-            "ANY virus (including CMV, EBV, HBV, HCV, HSV, VZV, or an unspecified virus), "
-            "or uses generic language such as 'donor is seropositive' without naming the virus. "
-            "Set SERONEGATIVE only if it explicitly states the donor was seronegative for all IgG tests. "
-            "Otherwise use NOT_MENTIONED."
-        ),
-    )
-
-class SerostatusDonorCMVMention(SpanAugmentedMention):
-    """
-    CMV serostatus of the donor at the time of first renal transplant.
-    """
-    serostatus: Serostatus = Field(
-        Serostatus.NOT_MENTIONED,
-        description=(
-            "CMV serostatus of the renal donor. Choose one of: 'SEROPOSITIVE', 'SERONEGATIVE', 'NOT_MENTIONED'. "
-            "Look for patterns like 'CMV D+', 'CMV D-', 'donor CMV IgG positive', 'donor CMV IgG negative', etc."
-        ),
-    )
-
-class SerostatusDonorEBVMention(SpanAugmentedMention):
-    """
-    EBV serostatus of the donor at the time of first renal transplant.
-    """
-    serostatus: Serostatus = Field(
-        Serostatus.NOT_MENTIONED,
-        description=(
-            "EBV serostatus of the renal donor. Choose one of: 'SEROPOSITIVE', 'SERONEGATIVE', 'NOT_MENTIONED'. "
-            "Look for patterns like 'EBV D+', 'EBV D-', 'donor EBV IgG positive', 'donor EBV IgG negative', etc."
-        ),
-    )
-
-class SerostatusRecipientMention(SpanAugmentedMention):
-    """
-    Overall serostatus of the recipient at the time of first renal transplant.
-    """
-    serostatus: Serostatus = Field(
-        Serostatus.NOT_MENTIONED,
-        description=(
-            "Serostatus of the recipient at the time of renal transplant for ANY virus. "
-            "Set SEROPOSITIVE if the note documents the recipient as seropositive for "
-            "ANY virus (including CMV, EBV, HBV, HCV, HSV, VZV, or an unspecified virus), "
-            "or if the text states 'recipient is seropositive' without naming the virus. "
-            "Set SERONEGATIVE only if explicitly stated the recipient was seronegative. "
-            "Otherwise use NOT_MENTIONED."
-        ),
-    )
-
-class SerostatusRecipientCMVMention(SpanAugmentedMention):
-    """
-    CMV serostatus of the recipient at the time of first renal transplant.
-    """
-    serostatus: Serostatus = Field(
-        Serostatus.NOT_MENTIONED,
-        description=(
-            "CMV serostatus of the recipient at the time of renal transplant. "
-            "Choose one of: 'SEROPOSITIVE', 'SERONEGATIVE', 'NOT_MENTIONED'. "
-            "Look for patterns like 'CMV R+', 'CMV R-', 'recipient CMV IgG positive', 'recipient CMV IgG negative', etc."
-        ),
-    )
-
-class SerostatusRecipientEBVMention(SpanAugmentedMention):
-    """
-    EBV serostatus of the recipient at the time of first renal transplant.
-    """
-    serostatus: Serostatus = Field(
-        Serostatus.NOT_MENTIONED,
-        description=(
-            "EBV serostatus of the recipient at the time of renal transplant. "
-            "Choose one of: 'SEROPOSITIVE', 'SERONEGATIVE', 'NOT_MENTIONED'. "
-            "Look for patterns like 'EBV R+', 'EBV R-', 'recipient EBV IgG positive', 'recipient EBV IgG negative', etc."
-        ),
-    )
-
-
-
-###############################################################################
-# Therapeutic Status Compliance
-###############################################################################
-class RxTherapeuticStatus(StrEnum):
-    THERAPEUTIC = "Immunosuppression levels are documented as therapeutic, adequate, or within target range."
-    SUB_THERAPEUTIC = 'Immunosuppression levels are documented as subtherapeutic, insufficient, or below target range.'
-    SUPRA_THERAPEUTIC = 'Immunosuppression levels are documented as supratherapeutic, above therapeutic level, or above target range.'
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class RxTherapeuticStatusMention(SpanAugmentedMention):
-    rx_therapeutic_status: RxTherapeuticStatus = Field(
-        RxTherapeuticStatus.NONE_OF_THE_ABOVE,
-        description='In the present encounter, what is the documented immunosuppression level?'
-    )
-
-
-###############################################################################
-# Medication Compliance
-###############################################################################
-class RxCompliance(StrEnum):
-    COMPLIANT = 'Patient is documented as compliant with immunosuppressive medications.'
-    PARTIALLY_COMPLIANT = "Patient is documented as only partially compliant with immunosuppressive medications."
-    NON_COMPLIANT = "Patient is documented as noncompliant with immunosuppressive medications."
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class RxComplianceMention(SpanAugmentedMention):
-    rx_compliance: RxCompliance = Field(
-        RxCompliance.NONE_OF_THE_ABOVE,
-        description='In the present encounter, is the patient documented as compliant with immunosuppressive medications?'
-    )
-
-
-###############################################################################
-# THE FOLLOWING DATA ELEMENTS TRACK BOTH
-# THE HISTORY AND THE PRESENT STATUS OF VARIABLES
-###############################################################################
-
-
-###############################################################################
-# DSA Donor Specific Antibody
-###############################################################################
-class DSAPresent(StrEnum):
-    """
-    Notice: DSA is strongly related to `GraftRejectionPresent`.
-
-    Treatment of DSA includes immunosuppressive drugs, IVIG, and plasmapheresis (PLEX).
-
-    Treatment with immunosuppressive drugs does *NOT* imply SUSPECTED DSA,
-    as many of immunosuppressive drugs are routinely used for "maintenance" therapy.
-
-    IVIG and plasmapheresis (PLEX) during the post-transplant (post induction) phase DOES imply
-    --> DSAPresent >  SUSPECTED (and possibly CONFIRMED)
-    --> `GraftRejectionPresent` > SUSPECTED (and possibly CONFIRMED or BIOPSY_PROVEN)
-    """
-    CONFIRMED = "DSA diagnostic test positive, DSA diagnosis 'confirmed' or 'positive', or increase in immunosuppression due to DSA"
-    SUSPECTED = "DSA suspected, DSA likely, DSA cannot be ruled out, DSA test result pending, or treatment with IVIG/plasmapheresis"
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class DSAMention(SpanAugmentedMention):
-    dsa_history: bool = Field(
-        False,
-        description="Does the patient have a past medical history of donor specific antibodies (DSA)?"
-    )
-    dsa: DSAPresent = Field(
-        DSAPresent.NONE_OF_THE_ABOVE,
-        description="What evidence documents donor specific antibodies (DSA) as current, active, or being evaluated/treated now?"
-    )
-
-
-############################################################################################################
-# Infection (** Any **)
-#   * necessary for PNA and UTI infections that often do not have a confirmed infection type!
-#   * useful as a secondary check to ensure more specific infection types are not missed
-############################################################################################################
-class InfectionPresent(StrEnum):
-    CONFIRMED = "Infection confirmed by laboratory test or imaging, infection diagnosis was 'confirmed' or 'positive', or reduced immunosuppression due to infection"
-    SUSPECTED = "Infection is suspected, likely, cannot be ruled out, infection is a differential diagnosis or infectious test result is pending"
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class InfectionMention(SpanAugmentedMention):
-    infection_history: bool = Field(
-        False,
-        description="Does the patient have a past medical history of an infection?"
-    )
-    infection: InfectionPresent = Field(
-        InfectionPresent.NONE_OF_THE_ABOVE,
-        description="What evidence documents infection as current, active, or being evaluated/treated now?"
-    )
-
-
-###############################################################################
-# Infection (Viral)
-###############################################################################
-class ViralInfectionPresent(StrEnum):
-    CONFIRMED = "Viral infection confirmed by laboratory test or imaging, viral infection diagnosis was 'confirmed' or 'positive', or reduced immunosuppression due to viral infection"
-    SUSPECTED = "Viral infection is suspected, likely, cannot be ruled out, viral infection is a differential diagnosis or viral test result is pending"
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class ViralInfectionMention(SpanAugmentedMention):
-    viral_infection_history: bool = Field(
-        False,
-        description="Does the patient have a past medical history of a viral infection?"
-    )
-    viral_infection: ViralInfectionPresent = Field(
-        ViralInfectionPresent.NONE_OF_THE_ABOVE,
-        description="What evidence documents viral infection as current, active, or being evaluated/treated now?"
-    )
-
-
-###############################################################################
-# Infection (Bacterial)
-###############################################################################
-class BacterialInfectionPresent(StrEnum):
-    CONFIRMED = "Bacterial infection confirmed by laboratory test or imaging, bacterial infection diagnosis was 'confirmed' or 'positive', or reduced immunosuppression due to bacterial infection"
-    SUSPECTED = "Bacterial infection is suspected, likely, cannot be ruled out, bacterial infection is a differential diagnosis or bacterial test result is pending"
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class BacterialInfectionMention(SpanAugmentedMention):
-    bacterial_infection_history: bool = Field(
-        False,
-        description="Does the patient have a past medical history of a bacterial infection?"
-    )
-    bacterial_infection: BacterialInfectionPresent = Field(
-        BacterialInfectionPresent.NONE_OF_THE_ABOVE,
-        description="What evidence documents bacterial infection as current, active, or being evaluated/treated now?"
-    )
-
-
-###############################################################################
-# Infection (Fungal)
-###############################################################################
-class FungalInfectionPresent(StrEnum):
-    CONFIRMED = "Fungal infection confirmed by laboratory test or imaging, fungal infection diagnosis was 'confirmed' or 'positive', or reduced immunosuppression due to fungal infection"
-    SUSPECTED = "Fungal infection is suspected, likely, cannot be ruled out, fungal infection is a differential diagnosis or fungal test result is pending"
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class FungalInfectionMention(SpanAugmentedMention):
-    fungal_infection_history: bool = Field(
-        False,
-        description="Does the patient have a past medical history of a fungal infection?"
-    )
-    fungal_infection: FungalInfectionPresent = Field(
-        FungalInfectionPresent.NONE_OF_THE_ABOVE,
-        description="What evidence documents fungal infection as current, active, or being evaluated/treated now?"
-    )
-
-
-###############################################################################
-# Graft Rejection
-###############################################################################
-class GraftRejectionPresent(StrEnum):
-    """
-    Notice: Graft rejection is strongly related to `DSAPresent`.
-
-    Treatment of graft rejection includes immunosuppressive drugs, IVIG, and plasmapheresis (PLEX).
-
-    Treatment with immunosuppressive drugs does *NOT* imply outcome is SUSPECTED,
-    as many of immunosuppressive drugs are routinely used for "maintenance" therapy.
-
-    IVIG and plasmapheresis (PLEX) during the post-transplant (post induction) phase DOES imply
-    --> `GraftRejectionPresent` > SUSPECTED (and possibly CONFIRMED or BIOPSY_PROVEN)
-    --> DSAPresent >  SUSPECTED (and possibly CONFIRMED)
-    """
-    BIOPSY_PROVEN = "Biopsy proven kidney graft rejection or pathology proven kidney graft rejection"
-    CONFIRMED = "Kidney graft rejection was 'diagnosed', 'confirmed' or 'positive'"
-    SUSPECTED = "Kidney graft rejection presumed, suspected, likely, cannot be ruled out, biopsy result pending, or treatment with IVIG/plasmapheresis"
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class GraftRejectionMention(SpanAugmentedMention):
-    graft_rejection_history: bool = Field(
-        False,
-        description="Does the patient have a past medical history of kidney graft rejection?"
-    )
-    graft_rejection: GraftRejectionPresent = Field(
-        GraftRejectionPresent.NONE_OF_THE_ABOVE,
-        description="What evidence documents kidney graft rejection as current, active, or being evaluated/treated now?"
-    )
-
-
-###############################################################################
-# Graft Failure
-###############################################################################
-class GraftFailurePresent(StrEnum):
-    CONFIRMED = "Kidney graft has failed or kidney graft loss"
-    SUSPECTED = "Kidney graft failure presumed, suspected, likely, or cannot be ruled out"
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class GraftFailureMention(SpanAugmentedMention):
-    graft_failure_history: bool = Field(
-        False,
-        description="Does the patient have a past medical history of kidney graft failure?"
-    )
-    graft_failure: GraftFailurePresent = Field(
-        GraftFailurePresent.NONE_OF_THE_ABOVE,
-        description="What evidence documents kidney graft failure as current, active, or being evaluated/treated now?"
-    )
-
-
-###############################################################################
-# PTLD
-###############################################################################
-class PTLDPresent(StrEnum):
-    """
-    Notice: PTLD treatments may also be used in 'rescue' therapy (DSA/graft rejection) or other cancers.
-    One notable difference from other cancers (such as skin cancer) is the absence of "surgical excision" (lymphoma).
-    """
-    BIOPSY_PROVEN = "Biopsy proven or pathology proven PTLD"
-    CONFIRMED = "PTLD was 'diagnosed', 'confirmed' or 'positive' or viral positive lymphoma"
-    SUSPECTED = "PTLD presumed, suspected, likely, cannot be ruled out, PTLD biopsy result pending, or treatment with chemotherapy/radiation"
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class PTLDMention(SpanAugmentedMention):
-    ptld_history: bool = Field(
-        False,
-        description="Does the patient have a past medical history of post transplant lymphoproliferative disorder (PTLD)?"
-    )
-    ptld: PTLDPresent = Field(
-        PTLDPresent.NONE_OF_THE_ABOVE,
-        description="What evidence documents post transplant lymphoproliferative disorder (PTLD) as current, active, or being evaluated/treated now?"
-    )
-
-
-###############################################################################
-# Cancer
-###############################################################################
-class CancerPresent(StrEnum):
-    """
-    Notice: Cancer treatments may also be used in 'rescue' therapy (DSA/graft rejection).
-    PTLD is a type of cancer. PTLD is a lymphoma and thus not treated with "surgical excision".
-    Skin cancer (of which there are many types carcinoma and melanoma) is treated with surgical excision.
-    """
-    BIOPSY_PROVEN = "Biopsy proven or pathology proven cancer"
-    CONFIRMED = "Cancer was 'diagnosed', 'confirmed' or 'positive'"
-    SUSPECTED = "Cancer is presumed, suspected, likely, cannot be ruled out, biopsy of any lesion, or treatment with chemotherapy/radiation"
-    NONE_OF_THE_ABOVE = "None of the above"
-
-
-class CancerMention(SpanAugmentedMention):
-    cancer_history: bool = Field(
-        False,
-        description="Does the patient have a past medical history of cancer?"
-    )
-    cancer: CancerPresent = Field(
-        CancerPresent.NONE_OF_THE_ABOVE,
-        description="What evidence documents cancer as current, active, or being evaluated/treated now?"
-    )
-
-
-###############################################################################
-# Deceased
-#  For tracking if the patient is noted to be deceased in any notes
-###############################################################################
-class DeceasedMention(SpanAugmentedMention):
-    deceased: bool | None = Field(
-        None,
-        description='Does the present encounter document that the patient is deceased?'
-    )
-    deceased_date: str | None = Field(
-        None,
-        description=(
-            'If the patient is deceased, include the date the patient became deceased. Use YYYY-MM-DD format if possible. '
-            'Use None if there is no date recorded or if the patient is not observed as deceased.'
-        )
-    )
-
-
-###############################################################################
-# Aggregated Annotation and Mention Classes
-###############################################################################
-
-class KidneyTransplantAnnotation(BaseModel):
-    """
-    An object-model for annotations of immune related adverse event (IRAE)
-    observations found in a patient's chart, relating specifically to kidney
-    transplants.
-    Take care to avoid false positives, like confusing information that only
-    appears in family history for patient history. Annotations should indicate
-    the relevant details of the finding, as well as some additional evidence
-    metadata to validate findings post-hoc.
-    """
-    donor_transplant_date_mention: DonorTransplantDateMention
-    donor_type_mention: DonorTypeMention
-    donor_relationship_mention: DonorRelationshipMention
-    donor_hla_match_quality_mention: DonorHlaMatchQualityMention
-    donor_hla_mismatch_count_mention: DonorHlaMismatchCountMention
-    donor_serostatus_mention: SerostatusDonorMention
-    donor_serostatus_cmv_mention: SerostatusDonorCMVMention
-    donor_serostatus_ebv_mention: SerostatusDonorEBVMention
-    recipient_serostatus_mention: SerostatusRecipientMention
-    recipient_serostatus_cmv_mention: SerostatusRecipientCMVMention
-    recipient_serostatus_ebv_mention: SerostatusRecipientEBVMention
-    rx_therapeutic_status_mention: RxTherapeuticStatusMention
-    rx_compliance_mention: RxComplianceMention
-    dsa_mention: DSAMention
-    infection_mention: InfectionMention
-    viral_infection_mention: ViralInfectionMention
-    bacterial_infection_mention: BacterialInfectionMention
-    fungal_infection_mention: FungalInfectionMention
-    graft_rejection_mention: GraftRejectionMention
-    graft_failure_mention: GraftFailureMention
-    ptld_mention: PTLDMention
-    cancer_mention: CancerMention
-    deceased_mention: DeceasedMention
-
-class MultipleTransplantHistoryAnnotation(BaseModel):
-    """
-    An object-model for annotations of patients with a history of multiple transplants.
-    Take care to avoid false positives, like confusing information that only
-    appears in family history for patient history. Annotations should indicate
-    the relevant details of the finding, as well as some additional evidence
-    metadata to validate findings post-hoc.
-    """
-    multiple_transplant_history_mention: MultipleTransplantHistoryMention
-
-
-class KidneyTransplantDonorGroupAnnotation(BaseModel):
-    """
-    An object-model for annotations of immune related adverse event (IRAE)
-    observations found in a patient's chart, relating specifically to kidney
-    transplants.
-    Take care to avoid false positives, like confusing information that only
-    appears in family history for patient history. Annotations should indicate
-    the relevant details of the finding, as well as some additional evidence
-    metadata to validate findings post-hoc.
-    """
-    donor_transplant_date_mention: DonorTransplantDateMention
-    donor_type_mention: DonorTypeMention
-    donor_relationship_mention: DonorRelationshipMention
-    donor_hla_match_quality_mention: DonorHlaMatchQualityMention
-    donor_hla_mismatch_count_mention: DonorHlaMismatchCountMention
-    donor_serostatus_mention: SerostatusDonorMention
-    donor_serostatus_cmv_mention: SerostatusDonorCMVMention
-    donor_serostatus_ebv_mention: SerostatusDonorEBVMention
-    recipient_serostatus_mention: SerostatusRecipientMention
-    recipient_serostatus_cmv_mention: SerostatusRecipientCMVMention
-    recipient_serostatus_ebv_mention: SerostatusRecipientEBVMention
-
-
-class KidneyTransplantLongitudinalAnnotation(BaseModel):
-    """
-    An object-model for annotations of immune related adverse event (IRAE)
-    observations found in a patient's chart, relating specifically to kidney
-    transplants.
-
-    This class only includes longitudinally variable mentions, i.e. those
-    that can change over time, such as therapeutic status, compliance, infections,
-    graft rejection/failure, DSA, PTLD, cancer, and deceased status.
-
-    Take care to avoid false positives, like confusing information that only
-    appears in family history for patient history. Annotations should indicate
-    the relevant details of the finding, as well as some additional evidence
-    metadata to validate findings post-hoc.
-    """
-    rx_therapeutic_status_mention: RxTherapeuticStatusMention
-    rx_compliance_mention: RxComplianceMention
-    dsa_mention: DSAMention
-    infection_mention: InfectionMention
-    viral_infection_mention: ViralInfectionMention
-    bacterial_infection_mention: BacterialInfectionMention
-    fungal_infection_mention: FungalInfectionMention
-    graft_rejection_mention: GraftRejectionMention
-    graft_failure_mention: GraftFailureMention
-    ptld_mention: PTLDMention
-    cancer_mention: CancerMention
-    deceased_mention: DeceasedMention
-
-
-class KidneyTransplantComplianceGroupAnnotation(BaseModel):
-    """
-    An object-model for annotations of immune related adverse event (IRAE)
-    observations found in a patient's chart, relating specifically to kidney
-    transplants. This class only includes therapeutic status and compliance mentions.
-    Take care to avoid false positives, like confusing information that only
-    appears in family history for patient history. Annotations should indicate
-    the relevant details of the finding, as well as some additional evidence
-    metadata to validate findings post-hoc.
-    """
-    rx_therapeutic_status_mention: RxTherapeuticStatusMention
-    rx_compliance_mention: RxComplianceMention
-
-
-class KidneyTransplantInfectionGroupAnnotation(BaseModel):
-    """
-    An object-model for annotations of immune related adverse event (IRAE)
-    observations found in a patient's chart, relating specifically to kidney
-    transplants. This class only includes infection and cancer-related mentions.
-
-    Take care to avoid false positives, like confusing information that only
-    appears in family history for patient history. Annotations should indicate
-    the relevant details of the finding, as well as some additional evidence
-    metadata to validate findings post-hoc.
-    """
-    infection_mention: InfectionMention
-    viral_infection_mention: ViralInfectionMention
-    bacterial_infection_mention: BacterialInfectionMention
-    fungal_infection_mention: FungalInfectionMention
-
-
-class KidneyTransplantRejectionFailureGroupAnnotation(BaseModel):
-    """
-    An object-model for annotations of immune related adverse event (IRAE)
-    observations found in a patient's chart, relating specifically to kidney
-    transplants. This class only includes graft rejection, DSA, graft failure mentions.
-
-    Take care to avoid false positives, like confusing information that only
-    appears in family history for patient history. Annotations should indicate
-    the relevant details of the finding, as well as some additional evidence
-    metadata to validate findings post-hoc.
-    """
-    dsa_mention: DSAMention
-    graft_rejection_mention: GraftRejectionMention
-    graft_failure_mention: GraftFailureMention
-
-
-class KidneyTransplantPostTransplantMalignanciesAnnotation(BaseModel):
-    """
-    An object-model for annotations of immune related adverse event (IRAE)
-    observations found in a patient's chart, relating specifically to kidney
-    transplants. This class only includes post transplant malignancies.
-
-    Take care to avoid false positives, like confusing information that only
-    appears in family history for patient history. Annotations should indicate
-    the relevant details of the finding, as well as some additional evidence
-    metadata to validate findings post-hoc.
-    """
-    ptld_mention: PTLDMention
-    cancer_mention: CancerMention
-
-
-class KidneyTransplantDeathGroupAnnotation(BaseModel):
-    """
-    An object-model for annotations of immune related adverse event (IRAE)
-    observations found in a patient's chart, relating specifically to kidney
-    transplants. This class only includes deceased mentions.
-
-    Take care to avoid false positives, like confusing information that only
-    appears in family history for patient history. Annotations should indicate
-    the relevant details of the finding, as well as some additional evidence
-    metadata to validate findings post-hoc.
-    """
-    deceased_mention: DeceasedMention
-
-
-###############################################################################
-# Artifacts for Label Studio usage
-###############################################################################
-
-# Enum describing all the relevant mention types' display labels
-# Names of Enum members should be 1 to 1 with the KidneyTransplantAnnotation fields
-class KidneyTransplantMentionLabels(StrEnum):
-    multiple_transplant_history_mention = auto()
-    donor_transplant_date_mention = auto()
-    donor_type_mention = auto()
-    donor_relationship_mention = auto()
-    donor_hla_match_quality_mention = auto()
-    donor_hla_mismatch_count_mention = auto()
-    donor_serostatus_mention  = auto()
-    donor_serostatus_cmv_mention = auto()
-    donor_serostatus_ebv_mention = auto()
-    recipient_serostatus_mention = auto()
-    recipient_serostatus_cmv_mention = auto()
-    recipient_serostatus_ebv_mention = auto()
-    rx_therapeutic_status_mention = auto()
-    rx_compliance_mention = auto()
-    dsa_mention = auto()
-    infection_mention = auto()
-    viral_infection_mention = auto()
-    bacterial_infection_mention = auto()
-    fungal_infection_mention = auto()
-    graft_rejection_mention = auto()
-    graft_failure_mention = auto()
-    ptld_mention = auto()
-    cancer_mention = auto()
-    deceased_mention = auto()
-
-
-#  Groups to be used in the eventual labelstudio interface
-# Right now this only encodes background, but could be used to organize/
-# determine sorting order, etc
-class KidneyTransplantMentionGroups(StrEnum):
-    MULTIPLE_TRANSPLANT_HISTORY = auto()
-    TRANSPLANT_DATE = auto()
-    DONOR = auto()
-    THERAPEUTIC = auto()
-    COMPLIANCE = auto()
-    DSA = auto()
-    INFECTION = auto()
-    REJECTION = auto()
-    CANCER = auto()
-    ENDPOINTS_FAILURE_DECEASED = auto()
-
-
-kidney_transplant_mention_groups_metadata = {
-    KidneyTransplantMentionGroups.MULTIPLE_TRANSPLANT_HISTORY: {
-        "background": "#FFD700"
-    },
-    KidneyTransplantMentionGroups.TRANSPLANT_DATE: {
-        "background": "#008B8B"
-    },
-    KidneyTransplantMentionGroups.DONOR: {
-        "background": "#00FFFF"
-    },
-    KidneyTransplantMentionGroups.THERAPEUTIC: {
-        "background": '#90ee90'
-    },
-    KidneyTransplantMentionGroups.COMPLIANCE: {
-        "background": '#20b2aa'
-    },
-    KidneyTransplantMentionGroups.DSA: {
-        "background": '#9370db'
-    },
-    KidneyTransplantMentionGroups.INFECTION: {
-        "background": '#DC143C'
-    },
-    KidneyTransplantMentionGroups.REJECTION: {
-        "background": "#FF00FF"
-    },
-    KidneyTransplantMentionGroups.CANCER: {
-        "background": "#FF8C00"
-    },
-    KidneyTransplantMentionGroups.ENDPOINTS_FAILURE_DECEASED: {
-        "background": "#00008B"
-    },
-}
-
-# Tying the Label Enum to the Group Enum and other metadata (display strings, e.g.)
-kidney_transplant_mention_ls_metadata = {
-    KidneyTransplantMentionLabels.multiple_transplant_history_mention: {
-        "display": "Multiple Transplant History",
-        "group": KidneyTransplantMentionGroups.TRANSPLANT_DATE,
-        "hotkey": "H",
-        "hotkey_mnemonic": "(H)istory of multiple transplants?",
-    },
-    KidneyTransplantMentionLabels.donor_transplant_date_mention: {
-        "display": "Transplant Date",
-        "group": KidneyTransplantMentionGroups.TRANSPLANT_DATE,
-        "hotkey": "D",
-        "hotkey_mnemonic": "(D)ate of first transplant",
-    },
-    KidneyTransplantMentionLabels.donor_type_mention: {
-        "display": "Donor Type",
-        "group": KidneyTransplantMentionGroups.DONOR,
-        "hotkey": "S",
-        "hotkey_mnemonic": "(S)urvival of donor?",
-    },
-    KidneyTransplantMentionLabels.donor_relationship_mention: {
-        "display": "Donor Relationship",
-        "group": KidneyTransplantMentionGroups.DONOR,
-        "hotkey": "K",
-        "hotkey_mnemonic": "(K)inship of donor?",
-    },
-    KidneyTransplantMentionLabels.donor_hla_match_quality_mention: {
-        "display": "Hla Match Quality",
-        "group": KidneyTransplantMentionGroups.DONOR,
-        "hotkey": "Q",
-        "hotkey_mnemonic": "(Q)uality HLA compatibility?",
-    },
-    KidneyTransplantMentionLabels.donor_hla_mismatch_count_mention: {
-        "display": "Hla Mismatch Count",
-        "group": KidneyTransplantMentionGroups.DONOR,
-        "hotkey": "M",
-        "hotkey_mnemonic": "(M)isMatch HLA compatibility?",
-    },
-    KidneyTransplantMentionLabels.donor_serostatus_mention: {
-        "display": "Donor serostatus +/-",
-        "group": KidneyTransplantMentionGroups.DONOR,
-        "hotkey": "O",
-        "hotkey_mnemonic": "(O)verall donor serostatus?",
-    },
-
-    KidneyTransplantMentionLabels.donor_serostatus_cmv_mention: {
-        "display": "Donor CMV serostatus",
-        "group": KidneyTransplantMentionGroups.DONOR,
-    },
-
-    KidneyTransplantMentionLabels.donor_serostatus_ebv_mention: {
-        "display": "Donor EBV serostatus",
-        "group": KidneyTransplantMentionGroups.DONOR,
-    },
-
-    KidneyTransplantMentionLabels.recipient_serostatus_mention: {
-        "display": "Recipient Serostatus +/-",
-        "group": KidneyTransplantMentionGroups.DONOR,
-        "hotkey": "X",
-        "hotkey_mnemonic": "e(X)isting serostatus of recipient",
-    },
-
-    KidneyTransplantMentionLabels.recipient_serostatus_cmv_mention: {
-        "display": "Recipient CMV serostatus",
-        "group": KidneyTransplantMentionGroups.DONOR,
-    },
-
-    KidneyTransplantMentionLabels.recipient_serostatus_ebv_mention: {
-        "display": "Recipient EBV serostatus",
-        "group": KidneyTransplantMentionGroups.DONOR,
-    },
-
-    KidneyTransplantMentionLabels.rx_therapeutic_status_mention: {
-        "display": "Rx Therapeutic",
-        "group": KidneyTransplantMentionGroups.THERAPEUTIC,
-        "hotkey": "T",
-        "hotkey_mnemonic": "(T)herapeutic Rx?",
-    },
-    KidneyTransplantMentionLabels.rx_compliance_mention: {
-        "display": "Rx Compliance",
-        "group": KidneyTransplantMentionGroups.COMPLIANCE,
-        "hotkey": "U",
-        "hotkey_mnemonic": "(U)sing Rx?",
-    },
-    KidneyTransplantMentionLabels.dsa_mention: {
-        "display": "DSA",
-        "group": KidneyTransplantMentionGroups.DSA,
-        "hotkey": "A",
-        "hotkey_mnemonic": "(A)ntibodies?",
-    },
-    KidneyTransplantMentionLabels.infection_mention: {
-        "display": "Infection",
-        "group": KidneyTransplantMentionGroups.INFECTION,
-        "hotkey": "I",
-        "hotkey_mnemonic": "(I)nfection?",
-    },
-    KidneyTransplantMentionLabels.viral_infection_mention: {
-        "display": "Viral",
-        "group": KidneyTransplantMentionGroups.INFECTION,
-        "hotkey": "V",
-        "hotkey_mnemonic": "(V)iral?",
-    },
-    KidneyTransplantMentionLabels.bacterial_infection_mention: {
-        "display": "Bacterial",
-        "group": KidneyTransplantMentionGroups.INFECTION,
-        "hotkey": "B",
-        "hotkey_mnemonic": "(B)acterial?",
-    },
-    KidneyTransplantMentionLabels.fungal_infection_mention: {
-        "display": "Fungal",
-        "group": KidneyTransplantMentionGroups.INFECTION,
-        "hotkey": "Z",
-        "hotkey_mnemonic": "Fungus like (Z)ombie ants",
-    },
-    KidneyTransplantMentionLabels.graft_rejection_mention: {
-        "display": "Graft Rejection",
-        "group": KidneyTransplantMentionGroups.REJECTION,
-        "hotkey": "R",
-        "hotkey_mnemonic": "Graft (R)ejection?",
-    },
-    KidneyTransplantMentionLabels.graft_failure_mention: {
-        "display": "Graft Failure",
-        "group": KidneyTransplantMentionGroups.ENDPOINTS_FAILURE_DECEASED,
-        "hotkey": "F",
-        "hotkey_mnemonic": "Graft (F)ailure?",
-    },
-    KidneyTransplantMentionLabels.ptld_mention: {
-        "display": "PTLD",
-        "group": KidneyTransplantMentionGroups.CANCER,
-        "hotkey": "P",
-        "hotkey_mnemonic": "PTLD?",
-    },
-    KidneyTransplantMentionLabels.cancer_mention: {
-        "display": "Cancer",
-        "group": KidneyTransplantMentionGroups.CANCER,
-        "hotkey": "C",
-        "hotkey_mnemonic": "(C)ancer?",
-    },
-    KidneyTransplantMentionLabels.deceased_mention: {
-        "display": "Deceased",
-        "group": KidneyTransplantMentionGroups.ENDPOINTS_FAILURE_DECEASED,
-        "hotkey": "X",
-        "hotkey_mnemonic": "(X) marks deceased",
-    },
-}
-
-
-if __name__ == "__main__":
-    basedir = os.path.dirname(__file__)
-
-    with open(f"{basedir}/donor.json", "w", encoding="utf8") as f:
-        json.dump(KidneyTransplantDonorGroupAnnotation.model_json_schema(), f, indent=2)
-
-    with open(f"{basedir}/history.json", "w", encoding="utf8") as f:
-        json.dump(MultipleTransplantHistoryAnnotation.model_json_schema(), f, indent=2)
-
-    with open(f"{basedir}/longitudinal.json", "w", encoding="utf8") as f:
-        json.dump(KidneyTransplantLongitudinalAnnotation.model_json_schema(), f, indent=2)
